requirements.txt
pyopenssl
rich
